+++ crates/storage/lib.rs
@@ -67,6 +67,7 @@
 pub mod api;
 pub mod backend;
 pub mod block_data_buffer;
+pub mod clean_cache;
 pub mod error;
 mod layering;
 pub mod migrations;